put everything in cart() in foreach loop to test session for product name

+++ public/cart.php
@@ -35,25 +35,33 @@ if(isset($_GET['delete'])) {
 }
 
 function cart() {
-    $query = query("SELECT * FROM products ");
-    confirm($query);
+
+    foreach ($_SESSION as $name => $value) {
+        if(substr($name, 0, 8) == "product_") {
+            
     
-    while($row = fetch_array($query)){
-        $product = <<<DELIMETER
-        <tr>
-            <td>{$row['product_title']}</td>
-            <td>&#36;{$row['product_price']}</td>
-            <td>{$row['product_quantity']}</td>
-            <td>2</td>
-            <td><a class="btn btn-warning" href="cart.php?remove={$row['product_id']}"><span class="glyphicon glyphicon-minus"></span></a>
-                <a class="btn btn-success" href="cart.php?add={$row['product_id']}"><span class="glyphicon glyphicon-plus"></span></a>
-                <a class="btn btn-danger"href="cart.php?delete={$row['product_id']}"><span class="glyphicon glyphicon-remove"></span></a></td>
-        </tr>
+            $query = query("SELECT * FROM products ");
+            confirm($query);
+    
+            while($row = fetch_array($query)){
+                $product = <<<DELIMETER
+                <tr>
+                    <td>{$row['product_title']}</td>
+                    <td>&#36;{$row['product_price']}</td>
+                    <td>{$row['product_quantity']}</td>
+                    <td>2</td>
+                    <td><a class="btn btn-warning" href="cart.php?remove={$row['product_id']}"><span class="glyphicon glyphicon-minus"></span></a>
+                        <a class="btn btn-success" href="cart.php?add={$row['product_id']}"><span class="glyphicon glyphicon-plus"></span></a>
+                        <a class="btn btn-danger"href="cart.php?delete={$row['product_id']}"><span class="glyphicon glyphicon-remove"></span></a></td>
+                </tr>
 DELIMETER;
         echo $product;
 
 
+            }
+        }
     }
 }
+
 ?>
